backend/requirements.txt
Flask
Flask-Cors
psycopg2-binary==2.9.9
python-dotenv
google-genai
Flask-Bcrypt
PyJWT
Flask-Limiter
anthropic
pypdf
gunicorn
sentry-sdk[flask]
